Implemented feedback from professor:
Updated handling of pasted text to display output directly on the web page instead of using popups.
Added helper text to guide users on how to use the application, based on the README file.
Improved user experience by switching to a monospace font for better readability and to prevent errors caused by invisible spaces.
Added comments in HTML and CSS files to mark sections and components, facilitating easier navigation and editing of the code.